Moneypool
tried to connect the rest adapter with
database
--> post endpoint for creating a new moneypool
--> get request for retrieving a moneypool

Problem: retrieving the right user information

+++ src/main/kotlin/com/wirvsvirus/solipool/moneypool/model/Moneypool.kt
@@ -1,6 +1,6 @@
 package com.wirvsvirus.solipool.moneypool.model
 
-import java.math.BigInteger
+import java.math.BigDecimal
 import java.util.*
 
-data class Moneypool(val id: UUID, val name: String, val creator: Int, val balance: BigInteger)
+data class Moneypool(val id: UUID, val name: String, val creator: Int, val balance: BigDecimal)
